Create ../tmp on first start

+++ bot/utils.py
@@ -18,14 +18,18 @@ def replaceChar(str):
 # preserving 60 pics in /tmp
 def cleanTmp():
     path = relativePathFix(os.path.join('..', 'tmp'))
-    pics = os.listdir(path)
-    if len(pics) >=59:
-        # sort pics with modified time
-        pics = sorted(pics, key = lambda x:os.path.getmtime(os.path.join(path, x)))
-    else:
+    if os.path.exists(path):
+        pics = os.listdir(path)
+        if len(pics) >=59:
+            # sort pics with modified time
+            pics = sorted(pics, key = lambda x:os.path.getmtime(os.path.join(path, x)))
+        else:
+            return
+        picsToDel = pics[0:len(pics)-59-1]
+        for pic in picsToDel:
+            os.remove(os.path.join(path, pic))
+        logging.info('Cleaned %d pics in tmp'%len(picsToDel))
         return
-    picsToDel = pics[0:len(pics)-59-1]
-    for pic in picsToDel:
-        os.remove(os.path.join(path, pic))
-    logging.info('Cleaned %d pics in tmp'%len(picsToDel))
-    return
+    else:
+        os.mkdir(path)
+        logging.info('mkdir ../tmp')
